Updated test wing
I added an extra 20x20 hole to the test wing. Then added another screw hole on the top of the test wing. I also made the the hole for the screw head a bit deeper so the screw head doesn't pop out of the wing when fully screwed. Additionally moved the screw holes to the center of the wing so it would look nicer and more organized.