Updated wing design
Here is the current wing that I have worked up too. I have moved the extrusions of the vertical 20x20 so that the bulk of the hole is in the middle wing while the right and left wings have a small amount of the extrusion. I am planning to polish up the inside art of the wing soon but the edges of the wings should be complete or close too.